Add "edit packet" mode
The user can now load a previously generated packet and do surgery
on it. It's a feature I've meant to implement for a long time but
only now it felt like I needed it badly, what with Jool's 232 issue
and all.

Well, the commit is really more than that, because it includes a
bunch of stuff I forgot to commit a long time ago. They concern the
"auto automatic" packet generator, which really doesn't matter.
(Those packets are already generated and even commited into Jool.
So I should probably just delete that code.)

The generator also spits a file called "src.txt" now, which is
standard input. This is so another packet can be generated quickly,
by editing src.txt according to needs and piping it to another
execution of the jar.

There might be other improvements, but I lost track.